Execute a reprex::reprex() and take a local screenshot of the reprex output

`reprex_shot()` will first take a `reprex` and then capture the HTML output into an on-disk image, optionally uploading the image to Imgur.

Usage

reprex_shot(filename = NULL, open_file = TRUE, imgur = FALSE, ...)

Arguments

filename

a filename, ending in .png

open_file

A logical, should the file be opened once saved

imgur

A logical, should the image be uploaded to imgur also

...

Arguments passed on to reprex::reprex

x

An expression. If not given, reprex() looks for code in input. If input is not provided, reprex() looks on the clipboard.

When the clipboard is structurally unavailable, e.g., on RStudio Server or RStudio Cloud, reprex() consults the current selection instead of the clipboard.

input

Character. If has length one and lacks a terminating newline, interpreted as the path to a file containing reprex code. Otherwise, assumed to hold reprex code as character vector. When input specifies a filepath, it also determines the reprex working directory and the location of all resulting files.

wd

An optional filepath that is consulted when input is not a filepath. (By default, all work is done, quietly, in a subdirectory of the session temp directory.)

The most common use of wd is to set wd = ".", which means "reprex right HERE in the current working directory". Do this if you really must demonstrate something with local files.

venue

Character. Must be one of the following (case insensitive):

"gh" for GitHub-Flavored Markdown, the default

"r" for a runnable R script, with commented output interleaved. Also useful for Slack code snippets; select "R" from the "Type" drop-down menu to enjoy nice syntax highlighting.

"rtf" for Rich Text Format (not supported for un-reprexing)

"html" for an HTML fragment suitable for inclusion in a larger HTML document (not supported for un-reprexing)

"slack" for pasting into a Slack message. Works best if you opt out of Slack's WYSIWYG interface and, instead, go to Preferences > Advanced and select "Format messages with markup".

"so" for Stack Overflow Markdown. Note: this is just an alias for "gh", since Stack Overflow started to support CommonMark-style fenced code blocks in January 2019.

"ds" for Discourse, e.g., community.rstudio.com. Note: this is currently just an alias for "gh".

render

Logical. Whether to call rmarkdown::render() on the templated reprex, i.e. whether to actually run the code. Defaults to TRUE. Exists primarily for the sake of internal testing.

advertise

Logical. Whether to include a footer that describes when and how the reprex was created. If unspecified, the option reprex.advertise is consulted and, if that is not defined, default is TRUE for venues "gh", "html", "so", "ds" and FALSE for "r", "rtf", "slack".

session_info

Logical. Whether to include sessioninfo::session_info(), if available, or sessionInfo() at the end of the reprex. When venue is "gh", the session info is wrapped in a collapsible details tag. Read more about opt().

style

Logical. Whether to set the knitr chunk option tidy = "styler", which re-styles code with the styler package. Read more about opt().

comment

Character. Prefix with which to comment out output, defaults to "#>". Read more about opt().

tidyverse_quiet

Logical. Sets the options tidyverse.quiet and tidymodels.quiet, which suppress (TRUE, the default) or include (FALSE) the startup messages for the tidyverse and tidymodels packages. Read more about opt().

std_out_err

Logical. Whether to append a section for output sent to stdout and stderr by the reprex rendering process. This can be necessary to reveal output if the reprex spawns child processes or system() calls. Note this cannot be properly interleaved with output from the main R process, nor is there any guarantee that the lines from standard output and standard error are in correct chronological order. See callr::r() for more. Read more about opt().

html_preview

Logical. Whether to show rendered output in a viewer (RStudio or browser). Always FALSE in a noninteractive session. Read more about opt().

outfile

wd or providing a filepath to input. To reprex in current working directory, use wd = "." now, instead of outfile = NA.

show

html_preview, for greater consistency with other R Markdown output formats.

si

session_info.

Value

a screenshot of the reprex on disk